requirements.txt
Django==4.2
djangorestframework==3.14
python-decouple==3.8
gunicorn==20.1
psycopg2-binary==2.9
